refactor(web): remove Phaser engine, swap home banner for a world snippet
- Replace the Phaser mascot clip on the homepage with a non-interactive PixiJS snippet of the /rooms world (GameWorld 'cover' fill mode fills the banner strip).
- Delete the entire Phaser engine (src/engine), the old /room route + views, and the mascot banner components; drop the phaser dependency.
- Decouple the Explore -> Rooms preset showcase from the engine via a static room-presets module (SVG preview now derives from plain dimensions).
