+++ website/package.json
@@ -40,7 +40,6 @@
 		"i18next-browser-languagedetector": "^8.2.0",
 		"i18next-http-backend": "^3.0.2",
 		"next": "^16.2.9",
-		"phaser": "^3.80.1",
 		"pixi.js": "^8.19.0",
 		"react": "^19.2.1",
 		"react-dom": "^19.2.1",
