--- website/src/engine/constants.ts
@@ -1,30 +1,0 @@
-export const ROOM_TILE_HEIGHT = 16;
-export const ROOM_TILE_WIDTH = 32;
-
-export const ROOM_WALL_L_OFFSET_X = -8;
-export const ROOM_WALL_L_OFFSET_Y = -123;
-export const ROOM_WALL_R_OFFSET_X = 64;
-export const ROOM_WALL_R_OFFSET_Y = -139;
-
-export const SELECTED_TILE_ASSET = "assets/selected_tile.png";
-export const SHADOW_TILE_ASSET = "assets/shadow_tile.png";
-
-export const AVATAR_SVG_WIDTH = 64;
-export const AVATAR_SVG_HEIGHT = 110;
-
-export const PRIORITY_DOOR_FLOOR = 1;
-export const PRIORITY_DOOR_FLOOR_SELECT = 2;
-export const PRIORITY_DOOR_FLOOR_PLAYER_SHADOW = 3;
-export const PRIORITY_DOOR_FLOOR_PLAYER = 4;
-export const PRIORITY_WALL = 6;
-export const PRIORITY_FLOOR = 7;
-export const PRIORITY_PLAYER_SHADOW = 8;
-export const PRIORITY_WALL_ITEM = 9;
-export const PRIORITY_FLOOR_SELECT = 11;
-export const PRIORITY_PLAYER = 11;
-export const PRIORITY_ROOM_ITEM = 11;
-export const PRIORITY_CHAT = 13;
-
-export const PRIORITY_MULTIPLIER = 10_000_000;
-export const COMPARABLE_X_Y = 1_000_000;
-export const COMPARABLE_Z = 10_000;
