--- website/src/engine/furniture/index.ts
@@ -1,7 +1,0 @@
-export { default as FurnitureRenderer } from "./FurnitureRenderer";
-export { FURNITURE_LAYOUTS } from "./layouts";
-export type {
-	FurnitureItemPlacement,
-	RoomFurnitureLayout,
-	FurnitureItemDefinition,
-} from "./types";
