--- website/src/engine/types.ts
@@ -1,11 +1,0 @@
-export type Direction = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7;
-
-export type AvatarAction = "idle" | "walking" | "waving" | "sitting";
-
-export type AvatarAppearance = {
-	skinColor: string;
-	hairColor: string;
-	hairStyle: number;
-	shirtColor: string;
-	pantsColor: string;
-};
