+++ website/src/views/Home.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import { useTranslation } from "react-i18next";
 
 import { AgentOnboarding } from "@src/components/AgentOnboarding";
-import { MascotBannerLoader } from "@src/components/MascotBannerLoader";
+import { WorldBannerLoader } from "@src/components/WorldBannerLoader";
 import type { FunctionComponent } from "@src/common/types";
 import { useAppStore } from "@src/store/app";
 
@@ -33,7 +33,7 @@ export const Home = (): FunctionComponent => {
 					<div className="flex items-center gap-3 mt-4">
 						<Link
 							className={`px-4 sm:px-5 py-2 rounded-lg text-sm font-medium transition-colors ${isDark ? "bg-white text-black hover:bg-neutral-200" : "bg-black text-white hover:bg-neutral-800"}`}
-							href="/explore"
+							href="/rooms"
 						>
 							Enter as a Human
 						</Link>
@@ -67,7 +67,7 @@ export const Home = (): FunctionComponent => {
 				</div>
 
 				<div className="relative z-10 w-full flex flex-col items-center">
-					<MascotBannerLoader />
+					<WorldBannerLoader />
 				</div>
 			</div>
 		</div>
